Expand user planet id limit, and cleanup id generator
uint16 was too short, it's easy to reach maximum, because of the removal
of planets.

---

Original commit (chain-contracts): bb36b73ea6c53542490e2c3640fbfcdc8cd21511

+++ chain-contracts/contracts/controllers/DebugController.sol
@@ -34,7 +34,7 @@ contract DebugController is UserGoldControllable, UserNormalPlanetControllable {
   }
 
   function debugMintMaxUserNormalPlanets(address account) external {
-    uint16 counter = userNormalPlanetIdCounterPermanence().read(account);
+    uint64 counter = userNormalPlanetIdCounterPermanence().read(account);
     require(userNormalPlanetRecordsCountOf(account) == 0, "you must not have planets");
 
     uint256[] memory arr = new uint256[](919);
